More review fixes
