+++ ydb/core/tablet_flat/ya.make
@@ -73,6 +73,7 @@ SRCS(
     flat_update_op.h
     probes.cpp
     shared_handle.cpp
+    shared_cache_counters.cpp
     shared_sausagecache.cpp
     shared_sausagecache.h
     tablet_flat_executor.h
